Add module your-rents-services-common-util, add dependency to your-rents-services-geodata, and refactor imports to use DataNotFoundException from common util package
