+++ pom.xml
@@ -13,6 +13,7 @@
 		<common.spring-boot.version>3.1.5</common.spring-boot.version>
 	</properties>
 	<modules>
+		<module>your-rents-services-common-util</module>
 		<module>your-rents-services-common-searchable</module>
 		<module>your-rents-services-geodata</module>
 	</modules>
