+++ your-rents-services-geodata/src/main/java/com/yourrents/services/geodata/repository/CountryRepository.java
@@ -5,8 +5,9 @@ import static com.yourrents.services.geodata.jooq.Tables.COUNTRY;
 import static org.jooq.Records.mapping;
 
 import com.yourrents.services.common.searchable.FilterCriteria;
+import com.yourrents.services.common.util.jooq.JooqUtils;
 import com.yourrents.services.geodata.model.Country;
-import com.yourrents.services.geodata.util.JooqUtils;
+
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
